supervisor: Keep cache sizes in off_t typed variables
This helps 32 bit systems, where ssize_t is just 32 bits.

Improves #161.

+++ src/firebuild/blob_cache.cc
@@ -345,7 +345,7 @@ int BlobCache::get_fd_for_file(const Hash &key) {
 void BlobCache::delete_entries(const std::string& path,
                                const std::vector<std::string>& entries,
                                const std::string& debug_postfix,
-                               ssize_t* debug_bytes) {
+                               off_t* debug_bytes) {
   struct stat st;
   for (const auto& entry : entries) {
     const std::string absolute_entry = path + "/" + entry;
@@ -379,8 +379,8 @@ void BlobCache::delete_entries(const std::string& path,
 
 void BlobCache::gc_blob_cache_dir(const std::string& path,
                                   const tsl::hopscotch_set<AsciiHash>& referenced_blobs,
-                                  ssize_t* cache_bytes, ssize_t* debug_bytes,
-                                  ssize_t* unexpected_file_bytes) {
+                                  off_t* cache_bytes, off_t* debug_bytes,
+                                  off_t* unexpected_file_bytes) {
   DIR * dir = opendir(path.c_str());
   if (dir == NULL) {
     return;
@@ -472,8 +472,8 @@ void BlobCache::gc_blob_cache_dir(const std::string& path,
   closedir(dir);
 }
 
-void BlobCache::gc(const tsl::hopscotch_set<AsciiHash>& referenced_blobs, ssize_t* cache_bytes,
-                   ssize_t* debug_bytes, ssize_t* unexpected_file_bytes) {
+void BlobCache::gc(const tsl::hopscotch_set<AsciiHash>& referenced_blobs, off_t* cache_bytes,
+                   off_t* debug_bytes, off_t* unexpected_file_bytes) {
   gc_blob_cache_dir(base_dir_, referenced_blobs, cache_bytes, debug_bytes, unexpected_file_bytes);
 }
